refact: api 코드 분리 후 위치 변경 #66
