+++ src/constants/constant.ts
@@ -23,3 +23,6 @@ export const ROUTER_PATH_REGEX = {
 
 // 플리 상세페이지에서 사용. 마지막 영상일 때, 다음 영상 제목으로 표시할 텍스트.
 export const LAST_VIDEO_TITLE = '플레이리스트 끝'
+
+// 플리 상세페이지에서 사용. 한번에 불러올 댓글 최대 개수
+export const COMMENT_CALL_LIMIT = 10
