+++ src/hooks/useDeleteComment.ts
@@ -1,14 +1,5 @@
+import deleteComment from '@/apis/comment/deleteComment'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-import axiosInstance from '@/apis/axiosInstance'
-import { DeleteCommentProps } from '@/types/comments'
-
-const deleteComment = async (props: DeleteCommentProps): Promise<void> => {
-  await axiosInstance.delete(`/comments`, {
-    params: {
-      comment_id: `eq.${props.commentId}`
-    }
-  })
-}
 
 const useDeleteComment = () => {
   const queryClient = useQueryClient()
