+++ src/hooks/useFetchUserLikeStatus.ts
@@ -1,19 +1,5 @@
+import fetchUserLikeStatus from '@/apis/like/fetchUserLikeStatus'
 import { useQuery } from '@tanstack/react-query'
-import axiosInstance from '@/apis/axiosInstance'
-
-const fetchUserLikeStatus = async (
-  userId: string,
-  playlistId: string
-): Promise<boolean> => {
-  const res = await axiosInstance.get('/likes', {
-    params: {
-      user_id: `eq.${userId}`,
-      playlist_id: `eq.${playlistId}`
-    }
-  })
-
-  return res.data.length > 0 // 데이터가 없는 경우 false 반환
-}
 
 /**
  * @param {string} userId - 유저 ID
